+++ .eslintrc
@@ -5,35 +5,36 @@
         "block-scoped-var": 2,
         "brace-style": 2,
         "callback-return": [2, ["callback", "cb", "next"]],
-        "comma-dangle": [ 2, "never" ],
+        "comma-dangle": [2, "never"],
         "comma-spacing": [2, {"before": false, "after": true}],
         "comma-style": [2, "last"],
-        "complexity": [ 1, 5 ],
+        "complexity": [1, 6],
         "computed-property-spacing": [2, "never"],
         "consistent-return": 2,
         "consistent-this": [2, "self"],
-        "curly": [ 2, "all" ],
+        "curly": [2, "all"],
         "default-case": 2,
         "dot-location": [2, "property"],
-        "dot-notation": [ 2, {"allowKeywords": true, "allowPattern": ""} ],
-        "eqeqeq": [ 2, "smart" ],
+        "dot-notation": [2, {"allowKeywords": true, "allowPattern": ""}],
+        "eqeqeq": [2, "smart"],
         "eol-last": 2,
         "func-names": 2,
         "func-style": [2, "declaration"],
         "global-require": 2,
         "guard-for-in": 2,
         "handle-callback-err": [2, "^.*(e|E)rr" ],
-        "indent": [ 2, 4 ],
+        "indent": [2, 4],
         "key-spacing": [2, {"beforeColon": false, "afterColon": true}],
-        "linebreak-style": [ 2, "unix" ],
+        "linebreak-style": [2, "unix"],
         "lines-around-comment": [2, { "beforeBlockComment": true, "allowBlockStart": true }],
+        "max-depth": [2, 5],
         "max-nested-callbacks": [2, 5],
         "max-len": [2, 80, 8, {"ignoreUrls": true}],
         "new-parens": 2,
         "newline-after-var": [2, "always"],
         "no-array-constructor": 2,
         "no-case-declarations": 2,
-        "no-cond-assign": [ 2, "always" ],
+        "no-cond-assign": [2, "always"],
         "no-console": 1,
         "no-constant-condition": 2,
         "no-dupe-args": 2,
@@ -91,13 +92,13 @@
         "no-warning-comments": [1, { "terms": ["todo", "fixme", "xxx"], "location": "start" }],
         "no-with": 2,
         "object-curly-spacing": [2, "never"],
-        "one-var": [ 2, "never" ],
+        "one-var": [2, "never"],
         "operator-linebreak": [2, "before"],
         "padded-blocks": [2, "never"],
         "radix": 2,
         "quote-props": [2, "as-needed"],
-        "quotes": [ 2, "single" ],
-        "semi": [ 2, "always" ],
+        "quotes": [2, "single"],
+        "semi": [2, "always"],
         "semi-spacing": 2,
         "sort-vars": [2, {"ignoreCase": true}],
         "space-after-keywords": 2,
