+++ lib/vm-agent.js
@@ -166,67 +166,12 @@ var COMPARISON_FIELDS = [
     'zone_state'
 ];
 
-/*
- * VMAPI has default values for some fields when the fields are not actually
- * in the VM objects in Moray. For some of those fields such as:
- *
- *  autoboot: null
- *  brand: null
- *  firewall_enabled: false
- *  nics: []
- *  platform_build_stamp: null
- *  resolvers: []
- *  server_uuid: null
- *  snapshots: []
- *  state: null
- *  tags: {}
- *  zone_state: null
- *  zonepath: null
- *
- * vmadm *also* always has these, so any time they're different we still want to
- * do an update. The ones VMAPI always includes that are *not* always in vmadm
- * objects are put into VMAPI_UNSET_FIELDS here so that we can detect their
- * default values and not force an update when the vmadm object does not have
- * these but the VMAPI object does and has the default value.
- *
- * If we did do that update, we'd be updating such a VM on every restart since
- * VMAPI would always give us an object with these properties even if the
- * correct object were in Moray.
- *
- * We have also excluded:
- *
- *  customer_metadata: {}
- *  internal_metadata: {}
- *
- * since these are not in the COMPARISON_FIELDS above. If they're added we'll
- * need to do a bit more complicated comparison since we'll have to treat {}
- * from VMAPI and not having either of these from vmadm as equivalent.
- *
- * TODO XXX: add a note in the VMAPI code that if the list is updated and more
- * fields are added, we come back here and update this too.
- */
-var VMAPI_UNSET_FIELDS = {
-    alias: null,
-    billing_id: null,
-    cpu_cap: null,
-    cpu_shares: null,
-    create_timestamp: null,
-    datasets: [],
-    destroyed: null,
-    last_modified: null,
-    limit_priv: null,
-    max_locked_memory: null,
-    max_lwps: null,
-    max_physical_memory: null,
-    max_swap: null,
-    owner_uuid: null,
-    quota: null,
-    ram: null,
-    zfs_filesystem: null,
-    zfs_io_priority: null,
-    zpool: null,
-    image_uuid: null
-};
+// XXX
+// PERIODIC FIELDS vs. ^^:
+//    removes boot_timestamp
+//    adds zfs_data_recsize
+//    adds zfs_root_recsize
+
 
 function VmAgent(options) {
     var self = this;
@@ -258,6 +203,7 @@ function VmAgent(options) {
     self.initializeProperties();
 }
 
+
 /*
  * This sets or resets the resettable properties on the VmAgent instance.
  */
@@ -286,6 +232,7 @@ VmAgent.prototype.initializeProperties = function initializeProperties() {
     self.ready = false;
     self.lastSeenVms = {};
 };
+
 
 VmAgent.prototype.updateVmapiVm = function updateVmapiVm(vm_uuid, callback) {
     var self = this;
@@ -382,6 +329,7 @@ VmAgent.prototype.updateVmapiVm = function updateVmapiVm(vm_uuid, callback) {
     });
 };
 
+
 /*
  * This is called when a VM has failed an update. It is responsible for waiting
  * for (and incrementing) the delay and then re-queuing the update if there's
@@ -424,6 +372,7 @@ function retryFailedUpdate(err, vm_uuid, callback) {
     callback();
 };
 
+
 /*
  * The fact that we use this queue serially has the additional advantage
  * currently of debouncing updates that are coming in frequently. For example if
@@ -451,6 +400,7 @@ VmAgent.prototype.queueVm = function queueVm(vm_uuid) {
         self.queue.push(vm_uuid);
     }
 };
+
 
 VmAgent.prototype.setupWatcher = function setupWatcher() {
     var self = this;
@@ -508,8 +458,8 @@ function makeComparable(vmobj, source) {
 
         if (vmobj.hasOwnProperty(field)) {
             if (source === 'vmapi'
-                && VMAPI_UNSET_FIELDS.hasOwnProperty(field)
-                && !diff(VMAPI_UNSET_FIELDS[field], vmobj[field])) {
+                && VMAPI.VMAPI_UNSET_FIELDS.hasOwnProperty(field)
+                && !diff(VMAPI.VMAPI_UNSET_FIELDS[field], vmobj[field])) {
                 /*
                  * VMAPI always includes some fields, even when unset. So we
                  * skip those here so that the comparison can work.
@@ -522,6 +472,7 @@ function makeComparable(vmobj, source) {
 
     return (newVmobj);
 }
+
 
 VmAgent.prototype.initialUpdate = function initialUpdate(callback) {
     var self = this;
@@ -704,6 +655,7 @@ VmAgent.prototype.initialUpdate = function initialUpdate(callback) {
     });
 };
 
+
 VmAgent.prototype.start = function start() {
     var self = this;
 
@@ -758,6 +710,7 @@ VmAgent.prototype.start = function start() {
     });
 };
 
+
 /*
  * This stops the running watchers, kills the queue and tries to ensure that
  * this VmAgent won't do more work and won't block the caller from shutting
@@ -770,5 +723,6 @@ VmAgent.prototype.stop = function stop() {
 
     self.initializeProperties();
 };
+
 
 module.exports = VmAgent;
