+++ lib/vm-watcher.js
@@ -84,53 +84,6 @@ var util = require('util');
 
 
 /*
- * To just detect whether there has been a change to a VM we don't need to load
- * all the fields each time. Any fields we don't load save some overhead in both
- * gathering and processing. Most updates will modify last_modified for the VM,
- * so mostly it is the fields that don't which we need to load. These include:
- *
- *  - datasets
- *  - disks (refreservation,etc)
- *  - indestructible_delegated
- *  - indestructible_zoneroot
- *  - pid
- *  - quota
- *  - snapshots
- *  - state
- *  - zfs_data_compression
- *  - zfs_data_recsize
- *  - zfs_root_compression
- *  - zfs_root_recsize
- *  - zone_state
- *  - zoneid
- *
- * So we grab all of those along with 'last_modified' and the 'brand' and 'uuid'
- * to identify the zone.
- *
- */
-var PERIODIC_FIELDS = [
-    'brand',
-    'datasets',
-    'disks',
-    'indestructible_delegated',
-    'indestructible_zoneroot',
-    'last_modified',
-    'pid',
-    'quota',
-    'snapshots',
-    'state',
-    'uuid',
-    'zfs_data_compression',
-    'zfs_data_recsize',
-    'zfs_root_compression',
-    'zfs_root_recsize',
-    'zone_state',
-    'zoneid'
-];
-var PERIODIC_INTERVAL = 5000; // ms
-
-
-/*
  * The VmWatcher will emit these events:
  *
  * - VmCreated:   A VM has been created
@@ -168,6 +121,9 @@ function VmWatcher(opts) {
         updateVm: self.newUpdateHandler('fs')
     });
 
+    // XXX If someone only wants last_modified, we can avoid starting up the
+    // periodic watcher?
+
     self.periodicWatcher = new PeriodicWatcher({
         log: opts.log,
         periodic_fields: opts.periodic_fields,
@@ -179,6 +135,7 @@ function VmWatcher(opts) {
     EventEmitter.call(this);
 }
 util.inherits(VmWatcher, EventEmitter);
+
 
 /*
  * Update knownVm with the properties from updateObj if knownVm has no
@@ -221,6 +178,7 @@ function updateIfNewer(vmUuid, log, knownVm, updateObj) {
     return (updated);
 }
 
+
 VmWatcher.prototype.dispatchCreate = // eslint-disable-line
 function dispatchCreate(vmUuid, updateVmobj, watcher) {
     var self = this;
@@ -251,6 +209,7 @@ function dispatchCreate(vmUuid, updateVmobj, watcher) {
         }, 'VM has changed');
     }
 };
+
 
 VmWatcher.prototype.dispatchModify = // eslint-disable-line
 function dispatchModify(vmUuid, updateVmobj, watcher) {
@@ -287,6 +246,7 @@ function dispatchModify(vmUuid, updateVmobj, watcher) {
     }
 };
 
+
 VmWatcher.prototype.dispatchDelete = // eslint-disable-line
 function dispatchDelete(vmUuid, updateVmobj, watcher) {
     var self = this;
@@ -310,6 +270,7 @@ function dispatchDelete(vmUuid, updateVmobj, watcher) {
     self.emit('VmDeleted', vmUuid, watcher);
 };
 
+
 VmWatcher.prototype.newUpdateHandler = function newUpdateHandler(watcher) {
     var self = this;
 
@@ -325,6 +286,7 @@ VmWatcher.prototype.newUpdateHandler = function newUpdateHandler(watcher) {
         }
     });
 };
+
 
 VmWatcher.prototype.start = function start() {
     var self = this;
@@ -344,5 +306,8 @@ VmWatcher.prototype.stop = function stop() {
 
     self.removeAllListeners();
 };
+
+// XXX TODO: export watchers lists of fields (PERODIC_FIELDS, etc)
+
 
 module.exports = VmWatcher;
