+++ lib/vmapi-client.js
@@ -17,6 +17,115 @@
 var restify = require('restify-clients');
 
 
+/*
+ * VMAPI has default values for some fields when the fields are not actually
+ * in the VM objects in Moray. For some of those fields such as:
+ *
+ *  autoboot: null
+ *  brand: null
+ *  firewall_enabled: false
+ *  nics: []
+ *  platform_build_stamp: null
+ *  resolvers: []
+ *  server_uuid: null
+ *  snapshots: []
+ *  state: null
+ *  tags: {}
+ *  zone_state: null
+ *  zonepath: null
+ *
+ * vmadm *also* always has these, so any time they're different we still want to
+ * do an update. The ones VMAPI always includes that are *not* always in vmadm
+ * objects are put into VMAPI_UNSET_FIELDS here so that we can detect their
+ * default values and not force an update when the vmadm object does not have
+ * these but the VMAPI object does and has the default value.
+ *
+ * If we did do that update, we'd be updating such a VM on every restart since
+ * VMAPI would always give us an object with these properties even if the
+ * correct object were in Moray.
+ *
+ * We have also excluded:
+ *
+ *  customer_metadata: {}
+ *  internal_metadata: {}
+ *
+ * since these are not in the COMPARISON_FIELDS above. If they're added we'll
+ * need to do a bit more complicated comparison since we'll have to treat {}
+ * from VMAPI and not having either of these from vmadm as equivalent.
+ *
+ * TODO XXX: add a note in the VMAPI code that if the list is updated and more
+ * fields are added, we come back here and update this too.
+ */
+var VMAPI_UNSET_FIELDS = {
+    alias: null,
+    billing_id: null,
+    cpu_cap: null,
+    cpu_shares: null,
+    create_timestamp: null,
+    datasets: [],
+    destroyed: null,
+    image_uuid: null,
+    last_modified: null,
+    limit_priv: null,
+    max_locked_memory: null,
+    max_lwps: null,
+    max_physical_memory: null,
+    max_swap: null,
+    owner_uuid: null,
+    quota: null,
+    ram: null,
+    zfs_filesystem: null,
+    zfs_io_priority: null,
+    zpool: null
+};
+
+// Fields VMAPI adds if not set
+var VMAPI_DEFAULT_FIELDS;
+
+
+// Builds and returns a new object which includes all the keys from both a and b
+function mergeObjs(a, b) {
+    var c = {};
+    var keyIdx;
+    var keys;
+
+    keys = Object.keys(a);
+    for (keyIdx = 0; keyIdx < keys.length; keyIdx++) {
+        c[keys[keyIdx]] = a[keys[keyIdx]];
+    }
+
+    keys = Object.keys(b);
+    for (keyIdx = 0; keyIdx < keys.length; keyIdx++) {
+        c[keys[keyIdx]] = b[keys[keyIdx]];
+    }
+
+    return (c);
+}
+
+
+/*
+ * This will be the list of all the default fields an object will have in VMAPI
+ * including both those that are added even if they're unset, and those that are
+ * always set.
+ */
+VMAPI_DEFAULT_FIELDS = mergeObjs(VMAPI_UNSET_FIELDS, {
+    autoboot: null,
+    brand: null,
+    customer_metadata: {},
+    firewall_enabled: false,
+    internal_metadata: {},
+    nics: [],
+    platform_buildstamp: null,
+    resolvers: null,
+    server_uuid: null,
+    snapshots: [],
+    state: null,
+    tags: {},
+    zone_state: null,
+    zonepath: null
+});
+
+
 function VmapiClient(options) {
     this.options = options;
     this.log = options.log;
@@ -89,5 +198,8 @@ VmapiClient.prototype.getVms = function getVms(server, callback) {
     });
 };
 
+
+VmapiClient.VMAPI_DEFAULT_FIELDS = VMAPI_DEFAULT_FIELDS;
+VmapiClient.VMAPI_UNSET_FIELDS = VMAPI_UNSET_FIELDS;
 
 module.exports = VmapiClient;
