+++ lib/watchers/fs-watcher.js
@@ -12,6 +12,7 @@ var assert = require('assert-plus');
 var fs = require('fs');
 var vasync = require('vasync');
 
+
 function FsWatcher(opts) {
     var self = this;
 
@@ -42,6 +43,7 @@ function FsWatcher(opts) {
     self.zonesWatcher = null;
 }
 
+
 function findVmUuidFromFilename(filename) {
     var match;
     var re;
@@ -60,6 +62,7 @@ function findVmUuidFromFilename(filename) {
 
     return (vm);
 }
+
 
 /*
  * Finds the last modified timestamp for a VM via /etc/zones/<uuid>.xml and
@@ -141,6 +144,7 @@ FsWatcher.prototype.getCurrentZones = function getCurrentZones(callback) {
     });
 };
 
+
 FsWatcher.prototype.getConfigTimestamp = // eslint-disable-line
 function getConfigTimestamp(vm, callback) {
     var self = this;
@@ -170,6 +174,7 @@ function getConfigTimestamp(vm, callback) {
         callback(err, latest);
     });
 };
+
 
 /*
  * For config we only watch for change events because ... XXX
@@ -240,6 +245,7 @@ FsWatcher.prototype.watchConfig = function watchConfig(vmUuid, callback) {
     }
     callback();
 };
+
 
 FsWatcher.prototype.processState = function processState(callback) {
     var self = this;
@@ -372,6 +378,7 @@ FsWatcher.prototype.processState = function processState(callback) {
     });
 };
 
+
 FsWatcher.prototype.start = function start() {
     var self = this;
 
@@ -391,6 +398,7 @@ FsWatcher.prototype.start = function start() {
         self.log.debug('initial update complete');
     });
 };
+
 
 FsWatcher.prototype.stop = function stop() {
     var self = this;
@@ -420,5 +428,6 @@ FsWatcher.prototype.stop = function stop() {
     self.isProcessing = false;
     self.lastSeenVms = null;
 };
+
 
 module.exports = FsWatcher;
