+++ lib/watchers/periodic-watcher.js
@@ -42,6 +42,11 @@ var vmadm = require('vmadm');
  *  - zone_state
  *  - zoneid
  *
+ * We've removed zoneid and pid from this list as these will always change only
+ * with an accompanying state change and high-resolution state change notices
+ * are required there's another watcher for that. Additionally 'zoneid' is not
+ * visible in VMAPI.
+ *
  * So we grab all of those along with 'last_modified' and the 'brand' and 'uuid'
  * to identify the zone.
  *
@@ -53,7 +58,6 @@ var PERIODIC_FIELDS = [
     'indestructible_delegated',
     'indestructible_zoneroot',
     'last_modified',
-    'pid',
     'quota',
     'snapshots',
     'state',
@@ -62,8 +66,7 @@ var PERIODIC_FIELDS = [
     'zfs_data_recsize',
     'zfs_root_compression',
     'zfs_root_recsize',
-    'zone_state',
-    'zoneid'
+    'zone_state'
 ];
 var PERIODIC_INTERVAL = 5000; // ms
 
@@ -106,6 +109,7 @@ function PeriodicWatcher(opts) {
         self.periodic_interval = PERIODIC_INTERVAL;
     }
 }
+
 
 /*
  * Takes two VM objects, returns true if the VMs match, false otherwise.
@@ -304,24 +308,11 @@ PeriodicWatcher.prototype.stop = function stop() {
     }
 };
 
+
 PeriodicWatcher.prototype.__testonly__ = {
     compareVms: compareVms,
     cmpVm: cmpVm
 };
-
-/*
- * If a caller somehow has missed a 'VmDeleted' event, they may have a list of
- * VMs that includes a VM we have already marked as deleted. They will not
- * receive another such event from us, so this function allows a caller to pass
- * in a list of VMs they have and we'll let them know.
- *
- * Hmmm... is it a good idea?
- *
- *  - what about unexpected create/modified?
- *  - how do we know their object contains all the fields we care about?
- *
- * XXX
- */
 
 
 module.exports = PeriodicWatcher;
