+++ tests/data.js
@@ -129,6 +129,7 @@ var minimalVmapiVm = {
     image_uuid: null
 };
 
+
 module.exports = {
     minimalVmapiVm: minimalVmapiVm,
     smartosPayloads: smartosPayloads
