+++ tests/test.PeriodicWatcher.js
@@ -58,12 +58,14 @@ function waitEvent(t, evt, vmUuid, eventIdx) {
     _waitEvent();
 }
 
+
 test('find SmartOS image', function _test(t) {
     common.testFindSmartosImage(t, function _findSmartosCb(latest) {
         smartosImageUUID = latest;
         t.end();
     });
 });
+
 
 test('load existing VMs', function _test(t) {
     var opts = {};
@@ -81,6 +83,7 @@ test('load existing VMs', function _test(t) {
         t.end();
     });
 });
+
 
 test('starting PeriodicWatcher', function _test(t) {
     function _onVmUpdate(vmUuid, updateType /* , updateObj */) {
@@ -105,6 +108,7 @@ test('starting PeriodicWatcher', function _test(t) {
     t.end();
 });
 
+
 test('create VM', function _test(t) {
     var eventIdx = events.length;
     var payload = {
@@ -128,6 +132,7 @@ test('create VM', function _test(t) {
     });
 });
 
+
 test('stop VM', function _test(t) {
     var eventIdx = events.length;
     var opts = {};
@@ -145,6 +150,7 @@ test('stop VM', function _test(t) {
         }
     });
 });
+
 
 test('start VM', function _test(t) {
     var eventIdx = events.length;
@@ -164,6 +170,7 @@ test('start VM', function _test(t) {
     });
 });
 
+
 test('modify quota using ZFS', function _test(t) {
     var eventIdx = events.length;
 
@@ -178,6 +185,7 @@ test('modify quota using ZFS', function _test(t) {
         }
     );
 });
+
 
 test('put metadata using mdata-put', function _test(t) {
     var eventIdx = events.length;
@@ -194,6 +202,7 @@ test('put metadata using mdata-put', function _test(t) {
         }
     );
 });
+
 
 test('delete VM', function _test(t) {
     var eventIdx = events.length;
@@ -212,11 +221,13 @@ test('delete VM', function _test(t) {
     });
 });
 
+
 test('stop PeriodicWatcher', function _test(t) {
     watcher.stop();
     t.ok(true, 'stopped watcher');
     t.end();
 });
+
 
 test('check SmartOS VM\'s events', function _test(t) {
     var evts = [];
