+++ tests/test.VmAgentMockVmadm.js
@@ -103,6 +103,7 @@ test('Startup VmAgent with VM missing from VMAPI', function _test(t) {
     vmAgent.start();
 });
 
+
 /*
  * Validate that when VmAgent starts up and vmadm lookup is missing a VM that
  * "GET /vms?state=active&server_uuid=..." included, that this missing VM is
@@ -139,6 +140,7 @@ test('Startup VmAgent with VM missing from vmadm', function _test(t) {
     vmAgent = new VmAgent(config);
     vmAgent.start();
 });
+
 
 /*
  * Start with vmapi + vmadm empty, then create some VMs. Then perform some
@@ -309,6 +311,7 @@ function _test(t) {
     _waitDone();
 });
 
+
 /*
  * When VmAgent starts, VMAPI is unavailable and there's a VM in vmadm that
  * does not exist in VMAPI. After 5 failed attempts (delta should be growing)
@@ -383,6 +386,7 @@ test('VmAgent retries when VMAPI returning errors', function _test(t) {
 
     _waitDone();
 });
+
 
 /*
  * VmAgent starts, there's a single VM in vmadm that gets updated to VMAPI.
@@ -525,6 +529,7 @@ test('VmAgent retries when VMAPI errors on PUT /vms/<uuid>', function _test(t) {
     _waitDone();
 });
 
+
 /*
  * VmAgent starts, there's a single VM in vmadm that gets updated to VMAPI.
  * After VMAPI is updated, it crashes and starts returning ECONNREFUSED, the VM
@@ -615,6 +620,7 @@ test('VmAgent sends deletion events after PUT failures', function _test(t) {
 
     _waitDone();
 });
+
 
 /*
  * vmadm has 2 VMs one which exists in VMAPI and one that doesn't. The one that
